[refactor] ErrorBoundary/AppRoutes 구조 정리
